requirements.txt
aioredis==1.3.1
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
backcall==0.1.0
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.1.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9
daphne==2.4.1
decorator==4.4.0
Django==3.0.6
django-admin-tools==0.9.0
django-automated-logging==5.0.1
django-baton==1.5.8
django-cleanup==5.1.0
django-cors-headers==3.1.1
django-filter==2.2.0
django-grappelli==2.14.1
django-health-check==3.12.1
django-jet==1.0.8
django-location-field==2.1.0
django-multiselectfield==0.1.12
django-nested-inline==0.4.2
django-ratings==0.3.7
django-related-admin==0.7.0
django-tagging==0.5.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-base64==2.0
fcm-django==0.3.4
flower==0.9.5
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.51.0
grpcio==1.33.2
gunicorn==20.0.4
hiredis==1.0.1
httplib2==0.17.3
humanize==3.1.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
install==1.3.4
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
kavenegar==1.1.2
kombu==4.6.8
Markdown==3.1.1
msgpack==0.6.2
oauth2client==1.5.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.8.0
prompt-toolkit==2.0.10
proto-plus==1.11.0
protobuf==3.11.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.7
Pygments==2.4.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
rsa==4.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
tornado==6.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.7
webpay-bahamta==0.1.1
websocket-client==0.57.0
zipp==3.1.0
zope.interface==5.0.2
